PREFIXESの共通化
