+++ app/src/root/App.tsx
@@ -7,7 +7,7 @@ import {
   TableContainer,
   Tbody,
 } from '@chakra-ui/react';
-import { PREFIXES, fetchActivity } from './utils/sparql';
+import { fetchActivity } from './utils/sparql';
 import { SelectActivity } from './components/SelectActivity';
 import { SelectScene } from './components/SelectScene';
 import { SelectCamera } from './components/SelectCamera';
@@ -15,6 +15,7 @@ import { SelectMedia } from './components/SelectMedia';
 import { ResultVideo } from './components/ResultVideo';
 import { ResultImage } from './components/ResultImage';
 import FloatingNavigationLink from '../common/components/FloatingNavigationLink';
+import { PREFIXES } from '../common/utils/sparql';
 
 function App() {
   const [activityList, setActivityList] = useState<Map<string, string[]>>(
